# energy-consumption-monitoring/docs/api-spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Energy Consumption Monitoring API",
        "version": "1.0.0",
        "description": "校园智慧后勤能耗监测系统后端 API 文档。"
    },
    "paths": {
...
        "/api/alarms/{id}/handle/": {
            "post": {
                "operationId": "api_alarms_handle_create",
                "description": "处理单条告警并记录处理信息。",
                "summary": "处理告警",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "标识此 alarm 的 唯一整数值。",
                        "required": true
                    }
                ],
                "tags": [
                    "api"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlarmHandle"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AlarmHandle"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AlarmHandle"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alarm"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/alarms/statistics/": {
            "get": {
                "operationId": "api_alarms_statistics_retrieve",
                "description": "返回告警总量、状态分布与设备 Top10。",
                "summary": "获取告警统计",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Alarm": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "device": {
                        "type": "integer"
                    },
                    "device_detail": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Device"
                            }
                        ],
                        "readOnly": true
                    },
                    "rule": {
                        "type": "integer",
                        "nullable": true,
                        "title": "Alarm rule"
                    },
                    "rule_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "alarm_type": {
                        "$ref": "#/components/schemas/AlarmTypeEnum"
                    },
                    "alarm_value": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,12}(?:\\.\\d{0,6})?$",
                        "nullable": true
                    },
                    "alarm_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "$ref": "#/components/schemas/AlarmStatusEnum"
                    },
                    "handler": {
                        "type": "integer",
                        "readOnly": true,
                        "nullable": true
                    },
                    "handler_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "handle_time": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "remark": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 500
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "alarm_time",
                    "alarm_type",
                    "created_at",
                    "device",
                    "device_detail",
                    "handle_time",
                    "handler",
                    "handler_name",
                    "id",
                    "rule_name",
                    "updated_at"
                ]
            },
            "AlarmHandle": {
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/AlarmHandleStatusEnum"
                            }
                        ],
                        "default": "PROCESSED"
                    },
                    "remark": {
                        "type": "string",
                        "maxLength": 500
                    }
                }
            },
            "AlarmHandleStatusEnum": {
                "enum": [
                    "PROCESSED",
                    "IGNORED"
                ],
                "type": "string",
                "description": "* `PROCESSED` - PROCESSED\n* `IGNORED` - IGNORED"
            },
...
            "AlarmStatusEnum": {
                "enum": [
                    "PENDING",
                    "PROCESSED",
                    "IGNORED"
                ],
                "type": "string",
                "description": "* `PENDING` - Pending\n* `PROCESSED` - Processed\n* `IGNORED` - Ignored"
            },
            "AlarmTypeEnum": {
                "enum": [
                    "THRESHOLD",
                    "MUTATION",
                    "OFFLINE"
                ],
                "type": "string",
                "description": "* `THRESHOLD` - Threshold\n* `MUTATION` - Mutation\n* `OFFLINE` - Offline"
            },
...
            "CodeEnum": {
                "enum": [
                    "WATER",
                    "ELECTRICITY",
                    "GAS"
                ],
                "type": "string",
                "description": "* `WATER` - Water\n* `ELECTRICITY` - Electricity\n* `GAS` - Gas"
            },
...
            "Device": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "device_id": {
                        "type": "string",
                        "maxLength": 64
                    },
                    "name": {
                        "type": "string",
                        "title": "Device name",
                        "maxLength": 128
                    },
                    "energy_type": {
                        "type": "integer"
                    },
                    "energy_type_detail": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/EnergyType"
                            }
                        ],
                        "readOnly": true
                    },
                    "room": {
                        "type": "integer",
                        "nullable": true
                    },
                    "room_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "floor_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "building_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "campus_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "model": {
                        "type": "string",
                        "nullable": true,
                        "title": "Device model",
                        "maxLength": 64
                    },
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StatusC1dEnum"
                            }
                        ],
                        "title": "Device status"
                    },
                    "last_data_time": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "building_name",
                    "campus_name",
                    "created_at",
                    "device_id",
                    "energy_type",
                    "energy_type_detail",
                    "floor_name",
                    "id",
                    "name",
                    "room_name",
                    "updated_at"
                ]
            },
...
            "EnergyType": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "title": "Energy name",
                        "maxLength": 64
                    },
                    "code": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/CodeEnum"
                            }
                        ],
                        "title": "Energy code"
                    },
                    "unit": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 128
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "code",
                    "created_at",
                    "id",
                    "name",
                    "unit",
                    "updated_at"
                ]
            },
...
            "StatusC1dEnum": {
                "enum": [
                    "ONLINE",
                    "OFFLINE",
                    "FAULT"
                ],
                "type": "string",
                "description": "* `ONLINE` - Online\n* `OFFLINE` - Offline\n* `FAULT` - Fault"
            },
...
        },
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}